feat(@dpc-sdp/ripple-tide-search): wip for location search refactor
