+++ packages/ripple-tide-search/composables/useTideSearch.ts
@@ -8,7 +8,8 @@ import {
   useRuntimeConfig,
   navigateTo,
   getSingleQueryStringValue,
-  scrollToElementTopWithOffset
+  scrollToElementTopWithOffset,
+  getScopedQueryParams
 } from '#imports'
 import type {
   TideSearchListingConfig,
@@ -699,19 +700,6 @@ export default ({
       }
 
       return acc
-    }, {})
-  }
-
-  /**
-   * Get a scoped set of query parameters
-   * i.e., custom location[] and search[] parameters
-   */
-  const getScopedQueryParams = (
-    scope: string,
-    params: { [key: string]: any }
-  ) => {
-    return Object.entries(params || {}).reduce((obj, [key, value]) => {
-      return { ...obj, [`${scope}[${key}]`]: value }
     }, {})
   }
